+++ emmet-api/emmet/api/core/__init__.py
@@ -1,0 +1,2 @@
+from emmet.api.core.api import MAPI
+from emmet.api.core.settings import MAPISettings